+++ common/src/gaussian.cpp
@@ -36,7 +36,9 @@
  */
 
 #include <pcl/common/gaussian.h>
+
 #include <cassert>
+#include <cstddef>
 
 void 
 pcl::GaussianKernel::compute (float sigma, 
@@ -147,7 +149,7 @@ pcl::GaussianKernel::convolveRows (const pcl::PointCloud<float>& input,
     {
       output.width = input.width;
       output.height = input.height;
-      output.resize (input.height * input.width);
+      output.resize (static_cast<std::size_t>(input.height * input.width));
     }
     unaliased_input = &input;
   }
@@ -191,7 +193,7 @@ pcl::GaussianKernel::convolveCols (const pcl::PointCloud<float>& input,
     {
       output.width = input.width;
       output.height = input.height;
-      output.resize (input.height * input.width);
+      output.resize (static_cast<std::size_t>(input.height * input.width));
     }
     unaliased_input = &input;
   }
